chore: move actions from utils to /actions
need to expand the 'actions/user' dir from index - do not like this pattern now
