+++ apps/dashboard/actions/stripe/stripe-invoice.ts
@@ -1,3 +1,4 @@
+'use server'
 import Stripe from 'stripe';
 
 export const createInvoice = async (
