+++ apps/dashboard/actions/stripe/stripe-products.ts
@@ -1,3 +1,4 @@
+'use server'
 import Stripe from 'stripe';
 
 const stripe = new Stripe(process.env.NEXT_PUBLIC_STRIPE_SECRET_KEY as string, {
