+++ apps/dashboard/actions/subscriptions/subscriptions.get.ts
@@ -1,3 +1,4 @@
+'use server'
 import { prisma } from '@/utils/prisma';
 
 export const getUserSubscription = async(userId: string) => {
