+++ apps/dashboard/app/(layowt)/page.tsx
@@ -1,6 +1,6 @@
 import { redirect } from 'next/navigation';
 import SiteLogo from '@/components/logo';
-import { getUserFromSession } from '@/utils/user';
+import { getUserFromSession } from '@/actions/user/index';
 
 export default async function App() {
   // get the user
